+++ tests/Http/Requests/RequestFactoryTest.php
@@ -269,7 +269,7 @@ JSON_API;
     {
         $this->expectedRecord = new stdClass();
         $this->adapter->method('exists')->with($resourceId)->willReturn(true);
-        $this->adapter->method('read')->with($resourceId)->willReturn($this->expectedRecord);
+        $this->adapter->method('find')->with($resourceId)->willReturn($this->expectedRecord);
 
         return $this;
     }