+++ CHANGELOG.md
@@ -15,6 +15,10 @@ All notable changes to this project will be documented in this file. This projec
 
 - [37](https://github.com/cloudcreativity/json-api/issues/37) Correctly detect message body in a HTTP response.
 
+### Deprecated
+
+- The JSON API request interface's `getRecord()` method has been deprecated.
+
 ## [0.11.0] - 2017-09-02
 
 ### Removed